src/requirements.txt
bitsandbytes==0.44.1
ipython==8.28.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
numpy==1.26.3
pandas==1.5.3
python-dotenv==1.0.1
pytorch-lightning==2.4.0
torch==2.5.1
torchaudio==2.5.0+cu121
torchmetrics==1.6.0
torchvision==0.20.0+cu121
transformers==4.45.2
umap-learn==0.5.7